swapped paid page into new page
